+++ com.unity.ml-agents/CHANGELOG.md
@@ -69,6 +69,7 @@ settings. Unfortunately, this may require retraining models if it changes the re
 or actuators on your system. (#5194)
 - Removed additional memory allocations that were occurring due to assert messages and iterating of DemonstrationRecorders. (#5246)
 - Fixed a bug where agent trying to access unintialized fields when creating a new RayPerceptionSensorComponent on an agent. (#5261)
+- Fixed a bug where the DemonstrationRecorder would throw a null reference exception if Num Steps To Record was > 0 and Record was turned off. (#5274)
 
 #### ml-agents / ml-agents-envs / gym-unity (Python)
 - Fixed a bug where --results-dir has no effect. (#5269)